third_party/blink/web_tests/platform/generic/editing/execCommand/arguments-combinations-expected.txt

These are tests for testing the how execCommand() works with different combinations of arguments. The "InsertHorizontalRule" command was chosen arbitrarily because it was what I was working on at the time, but the results should be paralleled in the other commands as well.

CONSOLE
PASS <hr> is <hr>
PASS <hr> is <hr>
PASS <hr id="foo"> is <hr id="foo">
PASS <hr id="foo"> is <hr id="foo">
PASS <hr> is <hr>
PASS <hr id="null"> is <hr id="null">
PASS <hr> is <hr>
PASS <hr id="0"> is <hr id="0">
PASS <hr> is <hr>
